clang-tidy fixes (#137)
* Tune .clang-tidy config and add test-specific overrides

- Disable bugprone-exception-escape and performance-enum-size (low
  value / conflicts with codebase's noexcept conventions)
- Add HeaderFilterRegex and configure readability-identifier-naming
  CheckOptions to match actual project conventions
- Add gtest/.clang-tidy and stress-tests/.clang-tidy inheriting the
  root config, disabling checks that are false positives against
  idiomatic test patterns (bugprone-use-after-move on moved-from
  state assertions, bugprone-unchecked-optional-access on the
  ASSERT_TRUE(x.has_value()) + x.value() gtest idiom, and
  bugprone-throwing-static-initialization on trivial static test
  fixture members)

* Fix all clang-tidy findings in src/

- lexer.hpp, parser.cpp: remove std::forward on template params reused
  across loop iterations (use-after-move pattern)
- ast.hpp, expression.hpp, options.hpp, table.hpp, test_results.hpp,
  token.hpp, context.hpp: add default member initializers
  (cppcoreguidelines-pro-type-member-init /
  clang-analyzer-optin.cplusplus.UninitializedObject)
- options.cpp/hpp: convert eager static-init lambda to a function-local
  static so exceptions during construction can propagate/be caught
  instead of guaranteed std::terminate (bugprone-throwing-static-initialization)
- table.cpp/hpp, get_args.hpp, registry.hpp: fix narrowing conversions
  in iterator arithmetic via explicit std::ptrdiff_t casts
- registry.hpp, util.hpp: replace string_view::data() usage with
  explicit std::string construction where a null-terminated C-string
  is required (real correctness risk on non-null-terminated slices);
  suppress the one legitimate pointer-range usage with a documented
  NOLINTNEXTLINE
- ast.cpp: pass vector by const reference instead of by value
  (performance-unnecessary-value-param)
- log.hpp, defines.hpp: rename Log() -> log_message() and
  CWT_CONCAT_ -> CWT_CONCAT_IMPL for identifier-naming conventions
- scanner.cpp: collapse branch-clone if/else, add explicit default
  case to character-dispatch switch
- catalog.cpp, report.cpp: cast indents to int before json::dump()

Verified with full rebuild + 555 unit tests + stress-tests + example
suite all passing. Full-codebase clang-tidy scan now reports zero
findings.

* Fix clang-tidy findings in gtest/

- scanner_spanish.cc: restore missing EXPECT_EQ assertion in
  spanish_keywords.feature_2 (was silently testing nothing,
  found via clang-analyzer-deadcode.DeadStores)
- scanner.cc: mark unused scan_token() result as [[maybe_unused]]
  in scan_vertical
- step_finder.cc: add default member initializers to test-local
  date struct

* Add clang-tidy CI workflow

Runs clang-tidy-17 on pull requests, scanning only files changed vs.
the PR base (src/gtest/stress-tests/examples, .cpp/.hpp/.cc). Since
main is protected and only reachable via merged PRs, every commit on
main has already passed this check, so a separate full-repo scan on
push isn't needed.

Diffing uses the correct three-dot merge-base semantics
(base.sha...HEAD) with --diff-filter=d to exclude deleted files,
fetching the base commit explicitly since PR checkouts are shallow
by default.

Findings are reported two ways:
- Inline GitHub Actions annotations (::warning file=...,line=...::)
  parsed from clang-tidy's output, shown directly on the PR diff
- Full raw log in $GITHUB_STEP_SUMMARY and uploaded as a build artifact

clang-tidy-17 isn't preinstalled on the ghcr.io/those1990/linux-dev:1.0
image (verified via a real CI run), so it's installed via apt-get.

Job fails if any finding is reported, since the codebase is currently
at zero clang-tidy findings.

* Bump outdated GitHub Actions to current major versions

actions/checkout v2/v3 -> v4 and actions/setup-python v2/v4 -> v5
across clang-tidy.yml, docs.yml, and unittests.yml. Addresses the
Node 20 deprecation warning surfaced by the new clang-tidy workflow
run.

+++ src/report.cpp
@@ -130,7 +130,7 @@ std::string as_json(std::size_t indents /* = 2 */)
     json_features.push_back(current_feature);
   }
 
-  return json_features.dump(indents);
+  return json_features.dump(static_cast<int>(indents));
 #else
   log::error(
       "nlohmann-json is not added as dependency in this build. Can not "
